requirements.txt
SQLAlchemy==2.0.31
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.25.2
python-multipart==0.0.20
email_validator==2.3.0
pydantic-settings==2.11.0
yfinance==0.2.40
requests==2.32.3
